+++ Assets/Prefabs/Loot/AttackerVirus.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 1689097345011191524}
   - component: {fileID: 2031234536257416508}
   - component: {fileID: 3032901014251215438}
+  - component: {fileID: -6327076564488574932}
   m_Layer: 0
   m_Name: AttackerVirus
   m_TagString: Untagged
@@ -120,3 +121,22 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   LootPool: {fileID: 0}
   lootType: 0
+--- !u!114 &-6327076564488574932
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8867622145172928444}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1e3fdca004f2d45fe8abbed571a8abd5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_OverrideArea: 0
+  m_Area: 0
+  m_OverrideGenerateLinks: 0
+  m_GenerateLinks: 0
+  m_IgnoreFromBuild: 1
+  m_ApplyToChildren: 1
+  m_AffectedAgents: ffffffff
