+++ Assets/ScriptableObjects/Spawner/AttackerVirus.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ef05f0931042faf4aa79d7a52c8f3c9e, type: 3}
+  m_Name: AttackerVirus
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+  lootType: 0
+  Weight: 2
