+++ Assets/ScriptableObjects/Spawner/AttackerVirus.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fd8ca3268eb55c44290a8eae9052f4d1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
