+++ Assets/ScriptableObjects/Spawner/CollectorVirus.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a2636ad3a5712eb4bbe10ee2232dd83c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
