+++ Assets/ScriptableObjects/Spawner/DamageBuff.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 339692942c00c6a448bf4ed917af4ee8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
