+++ Assets/ScriptableObjects/Spawner/DefenseBuff.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d726c6e8c5e2810419b3029382751401
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
