+++ Assets/ScriptableObjects/Spawner/HealerVirus.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b781e4aa1cdac748af62d0ea5d15fe5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
