+++ Assets/ScriptableObjects/Spawner/MovementBuff.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 81d2b7f79c1de5e4189ac8fc5d348c7b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
