+++ Assets/ScriptableObjects/Spawner/ScoreCollectable.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 38b02df01d757b142bee9db5d63d0ce6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
