+++ Assets/ScriptableObjects/Spawner/TankVirus.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8cdeb9c0a00419b4b95803eb33825ee8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
