+++ Assets/Scripts/Combat/Loot/LootSpawner.cs
@@ -1,12 +1,12 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using Lix.Core;
 
 public class LootSpawner : MonoBehaviour
 {
 
   // Define a list of weighted items
-  [SerializeField] private List<LootSpawnerItemSO> _weightedList = new List<LootSpawnerItemSO>();
 
   [Header("Spawn Settings")]
   [SerializeField] private float _spawnRate = 1f;
@@ -17,6 +17,7 @@ public class LootSpawner : MonoBehaviour
   private void Start()
   {
     StartCoroutine(SpawnLoop());
+    _lootPool = ServiceLocator.Get<LootPoolManager>();
   }
 
   private IEnumerator SpawnLoop()
@@ -36,7 +37,7 @@ public class LootSpawner : MonoBehaviour
     Transform spawnLocation = GetRandomSpawnLocation();
 
     // Select a weighted item at random
-    LootSpawnerItemSO item = SelectWeightedItem();
+    //LootSpawnerItemSO item = SelectWeightedItem();
 
     // Spawn the item
 
@@ -44,15 +45,15 @@ public class LootSpawner : MonoBehaviour
 
     // instantiate zombie
     Loot loot = _lootPool.GetLoot(lootType);
-    loot.transform.position = transform.position;
-    loot.transform.rotation = transform.rotation;
+    loot.transform.position = spawnLocation.position;
+    loot.transform.rotation = spawnLocation.rotation;
     loot.gameObject.SetActive(true);
 
     // Random scale
-    float randomScale = Random.Range(1f, _maxScaleMultiplier);
+    /*float randomScale = Random.Range(1f, _maxScaleMultiplier);
     Vector3 newScale = loot.transform.localScale;
     newScale *= randomScale;
-    loot.transform.localScale = newScale;
+    loot.transform.localScale = newScale;*/
   }
 
   private Transform GetRandomSpawnLocation()
@@ -60,31 +61,4 @@ public class LootSpawner : MonoBehaviour
     return transform.GetChild(Random.Range(0, transform.childCount));
   }
 
-  // Select a weighted item at random
-  private LootSpawnerItemSO SelectWeightedItem()
-  {
-    // Calculate the total weight of all items
-    float totalWeight = 0;
-    foreach (LootSpawnerItemSO item in _weightedList)
-    {
-      totalWeight += item.Weight;
-    }
-
-    // Generate a random number between 0 and the total weight
-    float randomWeight = Random.Range(0f, totalWeight);
-
-    // Loop through the items and subtract their weight from the random number
-    // until we find the selected item
-    foreach (LootSpawnerItemSO item in _weightedList)
-    {
-      if (randomWeight < item.Weight)
-      {
-        return item;
-      }
-      randomWeight -= item.Weight;
-    }
-
-    // This should never happen, but just in case
-    return null;
-  }
 }
